Updated MiniCoordinateWritable

+++ gila-common/src/unipg/gila/common/coordinatewritables/MiniCoordinateWritable.java
@@ -50,7 +50,7 @@ public class MiniCoordinateWritable implements Writable{
 	/**
 	 * The connected component index the vertex belongs to.
 	 */
-	protected long component;
+	protected int component;
 
 	public MiniCoordinateWritable() {
 		x = 0.0f;
@@ -58,13 +58,13 @@ public class MiniCoordinateWritable implements Writable{
 		component = -1;
 	}
 	
-	public MiniCoordinateWritable(float x, float y, long component){
+	public MiniCoordinateWritable(float x, float y, int component){
 		this.x = x;
 		this.y = y;
 		this.component = component;
 	}
 
-	public MiniCoordinateWritable(float x, float y, JSONArray oEs, long component) throws JSONException{
+	public MiniCoordinateWritable(float x, float y, JSONArray oEs, int component) throws JSONException{
 		this(x,y,component);
 		oneEdges = new LongWritableSet();
 		
@@ -96,7 +96,7 @@ public class MiniCoordinateWritable implements Writable{
 		return (Iterator<LongWritable>) oneEdges.iterator();
 	}
 		
-	public long getComponent() {
+	public int getComponent() {
 		return component;
 	}
 
@@ -107,7 +107,7 @@ public class MiniCoordinateWritable implements Writable{
 			oneEdges = new LongWritableSet();
 			oneEdges.readFields(in);
 		}
-		component = in.readLong();
+		component = in.readInt();
 	}
 
 	public void write(DataOutput out) throws IOException {
@@ -119,7 +119,7 @@ public class MiniCoordinateWritable implements Writable{
 			out.writeBoolean(true);
 			oneEdges.write(out);
 		}
-		out.writeLong(component);
+		out.writeInt(component);
 	}
 
 }
